+++ biopipen/scripts/scrna/SeuratMetadataMutater.R
@@ -1,4 +1,6 @@
+source("{{biopipen_dir}}/utils/misc.R")
 source("{{biopipen_dir}}/utils/mutate_helpers.R")
+
 library(rlang)
 library(tibble)
 library(dplyr)
@@ -14,7 +16,17 @@ metadata = srt@meta.data
 
 if (!is.null(metafile)) {
     mdata = read.table(metafile, header=TRUE, row.names=1, sep="\t", check.names=FALSE)
-    metadata = cbind(metadata, mdata[rownames(metadata),,drop=FALSE])
+    ov_cols = intersect(colnames(metadata), colnames(mdata))
+    if (length(ov_cols) > 0) {
+        log_warn(paste0(
+            "The following columns are already present in Seurat object and will be ignored: ",
+            paste(ov_cols, collapse=', ')
+        ))
+    }
+    metadata = cbind(
+        metadata,
+        mdata[rownames(metadata), setdiff(colnames(mdata), ov_cols), drop=FALSE]
+    )
 }
 
 expr = list()